Bug: Fix build error related to binding
