+++ app/src/main/java/com/mobileforce/hometeach/ui/withdrawalscreens/carddetails/CardDetailFragment.kt
@@ -9,7 +9,7 @@ import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.mobileforce.hometeach.R
 import com.mobileforce.hometeach.adapters.CircleTransform
-import com.mobileforce.hometeach.databinding.FragmentCarddetailsBinding
+import com.mobileforce.hometeach.databinding.FragmentCardDetailBinding
 import com.mobileforce.hometeach.ui.withdrawalscreens.MyCard
 import com.mobileforce.hometeach.ui.withdrawalscreens.MyCardModel
 import com.squareup.picasso.Picasso
@@ -24,25 +24,25 @@ private const val ARG_PARAM2 = "param2"
  * Use the [CarddetailsFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class CarddetailsFragment : Fragment() {
+class CardDetailFragment : Fragment() {
     // Rename and change types of parameters
-    private lateinit var card_list:MutableList<MyCard>
-    private lateinit var binding: FragmentCarddetailsBinding
+    private lateinit var cardList:MutableList<MyCard>
+    private lateinit var binding: FragmentCardDetailBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        binding = FragmentCarddetailsBinding.inflate(inflater, container, false)
+        binding = FragmentCardDetailBinding.inflate(inflater, container, false)
 
         return binding.root
-       // return inflater.inflate(R.layout.fragment_carddetails, container, false)
+       // return inflater.inflate(R.layout.fragment_card_detail, container, false)
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        card_list = mutableListOf()
-        card_list.add(
+        cardList = mutableListOf()
+        cardList.add(
             MyCard(
                 1,
                 R.drawable.ic_visa,
@@ -50,7 +50,7 @@ class CarddetailsFragment : Fragment() {
                 true
             )
         )
-        card_list.add(
+        cardList.add(
             MyCard(
                 2,
                 R.drawable.ic_master,
@@ -61,7 +61,7 @@ class CarddetailsFragment : Fragment() {
 
         var cards = MyCardModel(
             1,
-            card_list,
+            cardList,
             "Rahman Django",
             "profile_image",
             "215000 N"
