+++ app/src/main/res/layout/fragment_notification_tutor.xml
@@ -6,7 +6,7 @@
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".TutorDashBoardFragments.NotificationTutorFragment">
+        tools:context=".tutordashboardfragments.NotificationTutorFragment">
 
 
         <com.google.android.material.appbar.AppBarLayout
