+++ app/src/main/res/layout/list_item_class.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:background="@color/bg_main"
         android:layout_height="wrap_content"
-        tools:context=".TutorDashBoardFragments.OngoingFragment">
+        tools:context=".tutordashboardfragments.OngoingFragment">
 
         <LinearLayout
             android:id="@+id/classes"
@@ -23,7 +23,7 @@
             android:layout_gravity="center_horizontal"
             app:background="@color/bg_main"
             android:padding="10dp"
-            tools:context=".TutorDashBoardFragments.ClassesFragment">
+            tools:context=".tutordashboardfragments.ClassesFragment">
 
             <com.google.android.material.card.MaterialCardView
                 android:id="@+id/card"
